#10 Refactor codebase to introduce the steering microservice
Added the libcluon header file to establish communication among the microservices. Introduced the steering-related class with boilerplate code that accesses the shared memory. Revised both the Dockerfile and the CMakeLists.txt, so that to correspond to the changes made to the codebase as well as building and running the system accordingly.

Co-Authored-By: Krasen Parvanov <krasen@student.chalmers.se>
Co-Authored-By: Alexander Hultén <hultena@student.chalmers.se>
Co-Authored-By: Eemil Jeskanen <jeskanen@student.chalmers.se>